+++ src/sparseproduct.jl
@@ -1,33 +1,33 @@
-struct StaticProduct{NB}
+struct SparseProduct{NB}
    spec::Vector{NTuple{NB, Int}}
    # ---- temporaries & caches 
 end
 
-function StaticProduct()
-   return StaticProduct(bases, NTuple{NB, Int}[])
+function SparseProduct()
+   return SparseProduct(bases, NTuple{NB, Int}[])
 end
 
 # each column defines a basis element
-function StaticProduct(spec::Matrix{<: Integer})
+function SparseProduct(spec::Matrix{<: Integer})
    @assert all(spec .> 0)
    spect = [ Tuple(spec[:, i]...) for i = 1:size(spec, 2) ]
-   return StaticProduct(spect)
+   return SparseProduct(spect)
 end
  
-Base.length(basis::StaticProduct) = length(basis.spec)
+Base.length(basis::SparseProduct) = length(basis.spec)
 
 
 # ----------------------- evaluation interfaces 
 
 
-function evaluate(basis::StaticProduct, BB::Tuple{Vararg{AbstractVector}}) 
+function evaluate(basis::SparseProduct, BB::Tuple{Vararg{AbstractVector}}) 
    VT = mapreduce(eltype, promote_type, BB)
    A = zeros(VT, length(basis))
    evaluate!(A, basis, BB::Tuple)
    return A 
 end
 
-function evaluate(basis::StaticProduct, BB::Tuple{Vararg{AbstractMatrix}}) 
+function evaluate(basis::SparseProduct, BB::Tuple{Vararg{AbstractMatrix}}) 
    VT = mapreduce(eltype, promote_type, BB)
    nX = size(BB[1], 1)
    A = zeros(VT, nX, length(basis))
@@ -35,14 +35,14 @@ function evaluate(basis::StaticProduct, BB::Tuple{Vararg{AbstractMatrix}})
    return A 
 end
 
-test_evaluate(basis::StaticProduct, BB::Tuple) = 
+test_evaluate(basis::SparseProduct, BB::Tuple) = 
        [ prod(BB[j][basis.spec[i][j]] for j = 1:length(BB)) 
             for i = 1:length(basis) ]
    
 
 # ----------------------- evaluation kernels 
 
-function evaluate!(A, basis::StaticProduct{NB}, BB::Tuple{Vararg{AbstractVector}}) where {NB}
+function evaluate!(A, basis::SparseProduct{NB}, BB::Tuple{Vararg{AbstractVector}}) where {NB}
    @assert length(BB) == NB
    spec = basis.spec
    for (iA, ϕ) in enumerate(spec)
@@ -52,7 +52,7 @@ function evaluate!(A, basis::StaticProduct{NB}, BB::Tuple{Vararg{AbstractVector}
 end
 
 
-function evaluate!(A, basis::StaticProduct{NB}, BB::Tuple{Vararg{AbstractMatrix}}) where {NB}
+function evaluate!(A, basis::SparseProduct{NB}, BB::Tuple{Vararg{AbstractMatrix}}) where {NB}
    nX = size(BB[1], 1)
    @assert all(B->size(B, 1) == nX, BB)
    spec = basis.spec
@@ -67,13 +67,13 @@ end
 
 # -------------------- reverse mode gradient
 
-function _rrule_evaluate(basis::StaticProduct{NB}, BB::Tuple) where {NB}
+function _rrule_evaluate(basis::SparseProduct{NB}, BB::Tuple) where {NB}
    A = evaluate(basis, BB)
    return A, ∂A -> _pullback_evaluate(∂A, basis, BB)
 end
 
 
-function _pullback_evaluate(∂A, basis::StaticProduct{NB}, BB::Tuple) where {NB}
+function _pullback_evaluate(∂A, basis::SparseProduct{NB}, BB::Tuple) where {NB}
    TA = promote_type(eltype.(BB)...)
    ∂BB = ntuple(i -> zeros(TA, size(BB[i])...), NB)
    _pullback_evaluate!(∂BB, ∂A, basis, BB)
@@ -81,7 +81,7 @@ function _pullback_evaluate(∂A, basis::StaticProduct{NB}, BB::Tuple) where {NB
 end
 
 
-function _pullback_evaluate!(∂BB, ∂A, basis::StaticProduct{NB}, BB::Tuple) where {NB}
+function _pullback_evaluate!(∂BB, ∂A, basis::SparseProduct{NB}, BB::Tuple) where {NB}
    nX = size(BB[1], 1)
 
    @assert all(nX <= size(BB[i], 1) for i = 1:NB)
